% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_titanic.R
\name{taux_survie_class}
\alias{taux_survie_class}
\title{Taux de survie en fonction de la classe}
\usage{
taux_survie_class(data)
}
\arguments{
\item{data}{Data frame contenant les informations sur les passagers}
}
\value{
Un tableau avec le taux de survie pour chaque classe
}
\description{
Taux de survie en fonction de la classe
}
\examples{
taux_survie_class(df)
}
