% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_titanic.R
\name{taux_survie_sex}
\alias{taux_survie_sex}
\title{Taux de survie en fonction du sexe}
\usage{
taux_survie_sex(data)
}
\arguments{
\item{data}{Data frame contenant les informations sur les passagers}
}
\value{
Un tableau avec le taux de survie pour chaque sexe
}
\description{
Taux de survie en fonction du sexe
}
\examples{
taux_survie_sex(df)
}
\author{
Yamina Sadallah
}
